Compiled Python Script
This Script is same as above main.py the only difference this and that are this one is compiled so no one can see your source code and that one is not compiled so anyone can steel your code. You can try both of them by just run them both they both act as same and does same thing.